Example for Creating an inventory which will be used in the test to show on how the Multi tenancy works

+++ src/main/resources/db/migration/V1__initial_setup.sql
@@ -4,6 +4,6 @@ CREATE TABLE core.inventory
 (
     id uuid not null constraint pk_inventory primary key,
     name character varying not null,
-    amount numeric(19,2) not null,
+    amount int not null,
     tenant_id character varying not null
 );
